Create an outcome per path for the Greece
The goal of this story is to migrate the flow for
Greece to use the outcome per path strategy.

So far test artefacts haven't been affected by this change.

As you observe the title only contains opposite sex title, this
becuase the same sex title doesn't use the content title block.

It uses the markdown syntax, this will be fix in later commits