Update the content for what_sort_of_passport?
The description of citizen and alien passports is different for Estonia
and Latvia. Rather than create a duplicate question that routes to the
same place under the same conditions, but has slightly different copy in
the template I have decide to just add a little bit of logic in the
template to switch the copy.

Adding logic to the template is a temporary measure to get it working.
This should be refactored in a later commit.

No test artefacts were affected because the question text is only
generated once and it has already been created for the Estonia flow.