Add routing for Estonia in border_control question
If the user has:
1. selected Estonia in the first question
2. has an alien passport
3. is in transit

They should be directed to a new outcome. This outcome will implemented
in the next commit. For now, just the routing has been added to the flow
and the outcome Estonia goes to hasn’t changed.

Integration tests have been updated.
No test artefacts were affected by this change.