Rename Taiwan visa waiver outcome.
`outcome_taiwan_exception` has been renamed to match the visa exception 
outcome, `outcome_visit_waiver`. This means that it will be easier to 
find it in the flow and the logic documents.

Integration tests have been updated.

No test artefacts have been affected because will be corrected in the
following commit.