Support three questions country flow
This commit adds support for three questions country.

A three question country is one that only asks the user:

- Where they are getting married ?
- Where user currently resides ?
- What kind of marriage ? (i.e opposite or same sex)

The Partner Nationality question has been removed for Greece

Within methods have been defined and applied in the marriage abroad
calculator and flow that help identify a three questions country.

Based on the aforementioned the user is then shown relevant outcomes.

Some of the opposite/same sex outcomes are now redundant and have been
deleted. The existing six will be updated in later commit.