Preparatory: Fix transit route for DRC
This commit fixes the transit route for the democratic republic of the
congo.

This has been caused because of the change in the name and slug of the
aforementioned country appears to have been unintentionally omitted from
these PRs [1].

These PRs had been worked and merged around the same time and updating
the responses and expected-results were missed.

[1] #2559 #2785, #2778,