Re-open Iranian route service
This commit re-opens the Iranian route service for the register a death
flow.

This is achieved by removing Iran from the country_has_no_embassy list.

As a result of this removal, tests were required to be updated to mirror
the flow for Iran.

Regression test artefacts for the Iran route have been generated and
included in this commit.