Change first question for AdoptionCalculatorFlow
The phrasing and options for the first question in the
AdoptionCalculatorFlow need to change.

Changes just the output for first question and options in
AdoptionCalculatorFlow as a minimum, but subsequent changes will be
necessary to re-name the option variables and the question file
itself.

That work will be carried out in the next commits.