Apply 2 outcomes strategy to marriages in Egypt
This commits add Egypt to the list for two questions countries.

This reduces the number of question from 6 to 2. The number of outcomes
from 18 to 2 outcomes.

All regression test artefacts that applied to the 18 outcomes strategy,
have been removed and two regression tests artefacts added for opposite
and same sex only.

The responses and expected results yaml file has been updated, this is
due to the flow for Egypt has changed.